refactor scene data and it's related scripts
