+++ Assets/PetalsOfHope/Scenes/SampleScene.unity
@@ -119,54 +119,6 @@ NavMeshSettings:
     debug:
       m_Flags: 0
   m_NavMeshData: {fileID: 0}
---- !u!1 &78499287
-GameObject:
-  m_ObjectHideFlags: 0
-  m_CorrespondingSourceObject: {fileID: 0}
-  m_PrefabInstance: {fileID: 0}
-  m_PrefabAsset: {fileID: 0}
-  serializedVersion: 6
-  m_Component:
-  - component: {fileID: 78499289}
-  - component: {fileID: 78499288}
-  m_Layer: 0
-  m_Name: GameProgressionSystem
-  m_TagString: Untagged
-  m_Icon: {fileID: 0}
-  m_NavMeshLayer: 0
-  m_StaticEditorFlags: 0
-  m_IsActive: 1
---- !u!114 &78499288
-MonoBehaviour:
-  m_ObjectHideFlags: 0
-  m_CorrespondingSourceObject: {fileID: 0}
-  m_PrefabInstance: {fileID: 0}
-  m_PrefabAsset: {fileID: 0}
-  m_GameObject: {fileID: 78499287}
-  m_Enabled: 1
-  m_EditorHideFlags: 0
-  m_Script: {fileID: 11500000, guid: f89cf2be323758f43bc39782bf62f8b3, type: 3}
-  m_Name: 
-  m_EditorClassIdentifier: 
-  uniqueID: 67017d24-55ae-4c31-8be7-fd5023b19fb2
-  talismanDatabase: {fileID: 0}
-  onCollectibleCollected: {fileID: 0}
-  abilityCheckChannel: {fileID: 0}
---- !u!4 &78499289
-Transform:
-  m_ObjectHideFlags: 0
-  m_CorrespondingSourceObject: {fileID: 0}
-  m_PrefabInstance: {fileID: 0}
-  m_PrefabAsset: {fileID: 0}
-  m_GameObject: {fileID: 78499287}
-  serializedVersion: 2
-  m_LocalRotation: {x: 0, y: 0, z: 0, w: 1}
-  m_LocalPosition: {x: 0, y: 0, z: 0}
-  m_LocalScale: {x: 1, y: 1, z: 1}
-  m_ConstrainProportionsScale: 0
-  m_Children: []
-  m_Father: {fileID: 0}
-  m_LocalEulerAnglesHint: {x: 0, y: 0, z: 0}
 --- !u!4 &471915813 stripped
 Transform:
   m_CorrespondingSourceObject: {fileID: 378454640849913407, guid: def11a3d4ac5d1b4eacf68c629c6836e, type: 3}
@@ -440,6 +392,54 @@ Transform:
   m_Children: []
   m_Father: {fileID: 0}
   m_LocalEulerAnglesHint: {x: 0, y: 0, z: -90}
+--- !u!1 &797875015
+GameObject:
+  m_ObjectHideFlags: 0
+  m_CorrespondingSourceObject: {fileID: 0}
+  m_PrefabInstance: {fileID: 0}
+  m_PrefabAsset: {fileID: 0}
+  serializedVersion: 6
+  m_Component:
+  - component: {fileID: 797875017}
+  - component: {fileID: 797875016}
+  m_Layer: 0
+  m_Name: ProgressionSystemTester
+  m_TagString: Untagged
+  m_Icon: {fileID: 0}
+  m_NavMeshLayer: 0
+  m_StaticEditorFlags: 0
+  m_IsActive: 1
+--- !u!114 &797875016
+MonoBehaviour:
+  m_ObjectHideFlags: 0
+  m_CorrespondingSourceObject: {fileID: 0}
+  m_PrefabInstance: {fileID: 0}
+  m_PrefabAsset: {fileID: 0}
+  m_GameObject: {fileID: 797875015}
+  m_Enabled: 1
+  m_EditorHideFlags: 0
+  m_Script: {fileID: 11500000, guid: 98f236f94953e82488d4702797eb77ff, type: 3}
+  m_Name: 
+  m_EditorClassIdentifier: 
+  onCollectibleCollected: {fileID: 11400000, guid: e072762171ea83a4bb913b844a97edff, type: 2}
+  abilityCheckChannel: {fileID: 11400000, guid: 7a5fc313ed70e29428b60a6a3268e635, type: 2}
+  talismanToAward: {fileID: 11400000, guid: dc7b13cc75455bf4eaa1695e722b99c4, type: 2}
+  talismanIDToCheck: TALISMAN_DASH
+--- !u!4 &797875017
+Transform:
+  m_ObjectHideFlags: 0
+  m_CorrespondingSourceObject: {fileID: 0}
+  m_PrefabInstance: {fileID: 0}
+  m_PrefabAsset: {fileID: 0}
+  m_GameObject: {fileID: 797875015}
+  serializedVersion: 2
+  m_LocalRotation: {x: 0, y: 0, z: 0, w: 1}
+  m_LocalPosition: {x: 0, y: 0, z: 0}
+  m_LocalScale: {x: 1, y: 1, z: 1}
+  m_ConstrainProportionsScale: 0
+  m_Children: []
+  m_Father: {fileID: 0}
+  m_LocalEulerAnglesHint: {x: 0, y: 0, z: 0}
 --- !u!1001 &1355892530
 PrefabInstance:
   m_ObjectHideFlags: 0
@@ -777,6 +777,63 @@ Transform:
   m_Children: []
   m_Father: {fileID: 0}
   m_LocalEulerAnglesHint: {x: 0, y: 0, z: 0}
+--- !u!1001 &5525686792922782653
+PrefabInstance:
+  m_ObjectHideFlags: 0
+  serializedVersion: 2
+  m_Modification:
+    serializedVersion: 3
+    m_TransformParent: {fileID: 0}
+    m_Modifications:
+    - target: {fileID: 173996893479836642, guid: 596ea6d36dfaa8c4bb4ae64b67820341, type: 3}
+      propertyPath: m_Name
+      value: GameProgressionSystem
+      objectReference: {fileID: 0}
+    - target: {fileID: 472625676376002259, guid: 596ea6d36dfaa8c4bb4ae64b67820341, type: 3}
+      propertyPath: m_LocalPosition.x
+      value: 0
+      objectReference: {fileID: 0}
+    - target: {fileID: 472625676376002259, guid: 596ea6d36dfaa8c4bb4ae64b67820341, type: 3}
+      propertyPath: m_LocalPosition.y
+      value: 0
+      objectReference: {fileID: 0}
+    - target: {fileID: 472625676376002259, guid: 596ea6d36dfaa8c4bb4ae64b67820341, type: 3}
+      propertyPath: m_LocalPosition.z
+      value: 0
+      objectReference: {fileID: 0}
+    - target: {fileID: 472625676376002259, guid: 596ea6d36dfaa8c4bb4ae64b67820341, type: 3}
+      propertyPath: m_LocalRotation.w
+      value: 1
+      objectReference: {fileID: 0}
+    - target: {fileID: 472625676376002259, guid: 596ea6d36dfaa8c4bb4ae64b67820341, type: 3}
+      propertyPath: m_LocalRotation.x
+      value: 0
+      objectReference: {fileID: 0}
+    - target: {fileID: 472625676376002259, guid: 596ea6d36dfaa8c4bb4ae64b67820341, type: 3}
+      propertyPath: m_LocalRotation.y
+      value: 0
+      objectReference: {fileID: 0}
+    - target: {fileID: 472625676376002259, guid: 596ea6d36dfaa8c4bb4ae64b67820341, type: 3}
+      propertyPath: m_LocalRotation.z
+      value: 0
+      objectReference: {fileID: 0}
+    - target: {fileID: 472625676376002259, guid: 596ea6d36dfaa8c4bb4ae64b67820341, type: 3}
+      propertyPath: m_LocalEulerAnglesHint.x
+      value: 0
+      objectReference: {fileID: 0}
+    - target: {fileID: 472625676376002259, guid: 596ea6d36dfaa8c4bb4ae64b67820341, type: 3}
+      propertyPath: m_LocalEulerAnglesHint.y
+      value: 0
+      objectReference: {fileID: 0}
+    - target: {fileID: 472625676376002259, guid: 596ea6d36dfaa8c4bb4ae64b67820341, type: 3}
+      propertyPath: m_LocalEulerAnglesHint.z
+      value: 0
+      objectReference: {fileID: 0}
+    m_RemovedComponents: []
+    m_RemovedGameObjects: []
+    m_AddedGameObjects: []
+    m_AddedComponents: []
+  m_SourcePrefab: {fileID: 100100000, guid: 596ea6d36dfaa8c4bb4ae64b67820341, type: 3}
 --- !u!1001 &9211768950883067721
 PrefabInstance:
   m_ObjectHideFlags: 0
@@ -872,4 +929,5 @@ SceneRoots:
   - {fileID: 710886896}
   - {fileID: 9211768950883067721}
   - {fileID: 1355892530}
-  - {fileID: 78499289}
+  - {fileID: 5525686792922782653}
+  - {fileID: 797875017}
