--- Assets/PetalsOfHope/Scripts/Core/Data/SceneDataSO.cs.meta
@@ -1,2 +1,0 @@
-fileFormatVersion: 2
-guid: 54b7f6582eb87504d80c73a172a59f2d
