+++ Assets/PetalsOfHope/Scripts/Core/Events/LoadSceneRequestEventSO.cs
@@ -1,22 +1,22 @@
 using UnityEngine;
-
+using PetalsOfHope.Interfaces;
 
 namespace PetalsOfHope.Core.Events
 {
     [CreateAssetMenu(menuName = "Petals of Hope/Events/Load Scene Request Event")]
-    public class LoadSceneRequestEventSO : TypedEventSO<SceneDataSO>
+    public class LoadSceneRequestEventSO : TypedEventSO<ISceneData>
     {
-        // No additional implementation needed, inherits all functionality from TypedEventSO<SceneDataSO>
+        // This event uses the ISceneData interface as its payload.
+        // This decouples the Core event system from the concrete SceneDataSO class in the Data assembly.
         
         /*
         HOW TO USE:
-        1. Create an instance of this ScriptableObject in your project assets (e.g., via Right-Click -> Create -> Petals of Hope -> Events -> Load Scene Request Event).
-        2. Name it something like "LoadSceneRequestChannel".
-        3. In any script that needs to request a scene load (e.g., a UI button, a level-end trigger), add a SerializedField of this type:
-           [SerializeField] private LoadSceneRequestEventSO loadSceneRequestChannel;
-        4. Assign the "LoadSceneRequestChannel" asset to this field in the Inspector.
-        5. To trigger a scene load, call the Raise method with the SceneDataSO of the scene you want to load:
-           loadSceneRequestChannel.Raise(targetSceneData);
+        1. Create an instance of this ScriptableObject asset.
+        2. Name it "LoadSceneRequestChannel".
+        3. In a requester script (e.g., a UI button), get a reference to this channel.
+        4. In the requester, also get a reference to the specific SceneDataSO you want to load.
+        5. Call the Raise method, passing the SceneDataSO. It will be implicitly cast to ISceneData.
+           e.g., loadSceneRequestChannel.Raise(level1SceneData);
         */
     }
 }
