+++ Assets/PetalsOfHope/Scripts/Data/Stats/Character/CharacterStatsSo.cs
@@ -1,5 +1,5 @@
 using UnityEngine;
-using PetalsOfHope.Core.Data.Stats;
+using PetalsOfHope.Data.Stats;
 
 namespace PetalsOfHope.Data.Player
 {
