+++ Assets/PetalsOfHope/Scripts/Gameplay/Player/PlayerHealth.cs
@@ -3,7 +3,7 @@ using UnityEngine;
 using PetalsOfHope.Data.Player;
 using PetalsOfHope.Core.Events;
 using PetalsOfHope.Interfaces;
-using CoreAnimation = PetalsOfHope.Core.Animation.AnimationController; // Alias
+using CoreAnimation = PetalsOfHope.Gameplay.Animation.AnimationController; // Alias
 
 // In Assets/_Project/Scripts/Gameplay/Player/PlayerHealth.cs
 namespace PetalsOfHope.Gameplay.Player
