+++ Assets/PetalsOfHope/Scripts/Interfaces/ISceneData.cs.meta
@@ -1,0 +1,2 @@
+fileFormatVersion: 2
+guid: 7a5be6ecb90f5fb4a92c7f4080553f1a
